フォルダ構成の変更

+++ backend/functions/calc_average.py
@@ -2,12 +2,12 @@
 #ある時限の各値の平均が入ったリストを返す
 #[CO2平均，Temp平均，humi平均]
 def get_ave(data: list):
-    elements = [0,0,0]
+    elements = [0.0,0.0,0.0]
     wari = len(data)
     for ele in data:
-        elements[0] += int(ele[3])
-        elements[1] += int(ele[4])
-        elements[2] += int(ele[5])
+        elements[0] += float(ele[3])
+        elements[1] += float(ele[4])
+        elements[2] += float(ele[5])
     if(wari == 0):
         return
     elements[0] //= wari
